# docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_transport_http.AuthResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_at:
        example: "2025-11-12T15:04:05Z"
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/internal_transport_http.UserDTO'
    type: object
  internal_transport_http.ErrorResponse:
    properties:
      code:
        example: AUTH_001
        type: string
      error:
        example: invalid_credentials
        type: string
      message:
        example: Invalid email or password
        type: string
    type: object
  internal_transport_http.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: SecurePass123!
        type: string
    type: object
  internal_transport_http.LogoutRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refresh_token
    type: object
  internal_transport_http.MessageResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
  internal_transport_http.RefreshTokenRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  internal_transport_http.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      first_name:
        example: John
        type: string
      last_name:
        example: Doe
        type: string
      password:
        example: SecurePass123!
        minLength: 8
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  internal_transport_http.UpdateProfileRequest:
    properties:
      first_name:
        example: John
        type: string
      last_name:
        example: Doe
        type: string
    required:
    - first_name
    - last_name
    type: object
  internal_transport_http.UserDTO:
    properties:
      created_at:
        example: "2025-11-12T10:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      first_name:
        example: John
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      kyc_status:
        example: approved
        type: string
      last_name:
        example: Doe
        type: string
      updated_at:
        example: "2025-11-12T10:00:00Z"
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@pandora-exchange.com
    name: API Support
    url: https://pandora-exchange.com/support
  description: User authentication and management service for Pandora Exchange
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://pandora-exchange.com/terms
  title: Pandora Exchange User Service API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return access tokens
      parameters:
      - description: User login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/internal_transport_http.AuthResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
      summary: Login user
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a refresh token for a new access token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            $ref: '#/definitions/internal_transport_http.AuthResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
      summary: Refresh access token
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account and return authentication tokens
      parameters:
      - description: User registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/internal_transport_http.AuthResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
      summary: Register a new user
      tags:
      - Authentication
  /users/me:
    get:
      consumes:
      - application/json
      description: Get current user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            $ref: '#/definitions/internal_transport_http.UserDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update current user's profile information
      parameters:
      - description: Profile update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user profile
          schema:
            $ref: '#/definitions/internal_transport_http.UserDTO'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Users
  /users/me/logout:
    post:
      consumes:
      - application/json
      description: Invalidate refresh token and logout user from current device
      parameters:
      - description: Refresh token to invalidate
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/internal_transport_http.MessageResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - Authentication
  /users/me/logout-all:
    post:
      consumes:
      - application/json
      description: Invalidate all refresh tokens for the user
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/internal_transport_http.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout from all devices
      tags:
      - Authentication
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"